Implemented a "historyPoint" system into the stamps.
This can be used to tell stamps what point in the history they are in, which can be used together with the new editorUndo function, which tells stamps that the user tried undoing something.
Together with the historyPoints the stamps can figure out if they need to reset their stats for example.