Changing the way we store versions
Versions are now stored in a buildinfo.cfg config file, which is not on github as it is generated before built.
This allows us to record built time, type of build (release/dev/none) and possibly more info in the future.
Some changes had to be made so that it would work, but using SnipSniper with it is basicly the same experience.